+++ MOS-6502-Emulator/Instructions/LDY.swift
@@ -53,7 +53,7 @@ public class LDY
     public class func zeroPage( cpu: CPU ) throws
     {
         let address     = try cpu.readUInt8FromMemoryAtPC()
-        cpu.registers.Y = try cpu.readUInt8FromMemory( at: UInt64( address ) )
+        cpu.registers.Y = try cpu.readUInt8FromMemory( at: UInt16( address ) )
 
         LD.setStatus( for: cpu.registers.A, cpu: cpu )
     }
